add new favorites files structure
